[✏️Comment] ViewModel 명칭 통합 및 정리
