--- PalangPalang/Sources/Presenters/ViewModel/SettingsViewModel.swift
@@ -1,9 +1,0 @@
-//
-//  SettingsViewModel.swift
-//  PalangPalang
-//
-//  Created by 박혜운 on 8/2/24.
-//  Copyright © 2024 com.mc3. All rights reserved.
-//
-
-import Foundation
